+++ funcMath.h
@@ -3,4 +3,4 @@ float add(float number1, float number2);
 float subtract(float number1, float number2);
 float divide(float number1, float number2);
 float multiply(float number1, float number2);
-float factorial(float number);
+long factorial(long number);
